+++ backend/canisters/sns_neuron_controller/impl/src/jobs/mod.rs
@@ -1,5 +1,7 @@
-pub mod process_neurons;
+pub mod process_nns_neurons;
+pub mod process_sns_neurons;
 
 pub(crate) fn start() {
-    process_neurons::start_job();
+    process_sns_neurons::start_job();
+    process_nns_neurons::start_job();
 }