+++ backend/canisters/sns_neuron_controller/impl/src/state.rs
@@ -1,13 +1,19 @@
-use crate::types::neuron_manager::NeuronManager;
-use crate::types::neuron_manager::Neurons;
-use crate::types::neuron_metrics::NeuronWithMetric;
+use crate::types::icp_neuron_manager::IcpManager;
+use crate::types::neurons::nns_neurons::NnsNeuronWithMetric;
+use crate::types::neurons::sns_neurons::SnsNeuronWithMetric;
+use crate::types::sns_neuron_manager::NeuronManager;
+use crate::types::sns_neuron_manager::NeuronManagerEnum;
+use crate::types::GoldaoManager;
 use crate::types::OgyManager;
 use candid::{CandidType, Principal};
 use canister_state_macros::canister_state;
 use serde::{Deserialize, Serialize};
 use sns_governance_canister::types::Neuron;
+use sns_neuron_controller_api_canister::init::GoldaoManagerConfig;
+use sns_neuron_controller_api_canister::init::IcpManagerConfig;
+use sns_neuron_controller_api_canister::init::OgyManagerConfig;
+use sns_neuron_controller_api_canister::neuron_type::NeuronType;
 use types::BuildVersion;
-use types::CanisterId;
 use types::TimestampMillis;
 use utils::{
     env::{CanisterEnv, Environment},
@@ -40,6 +46,8 @@ impl RuntimeState {
             authorized_principals: self.data.authorized_principals.clone(),
             rewards_destination: self.data.rewards_destination,
             ogy_neuron_manager_metrics: self.data.neuron_managers.ogy.get_neuron_metrics(),
+            goldao_neuron_manager_metrics: self.data.neuron_managers.goldao.get_neuron_metrics(),
+            icp_neuron_manager_metrics: self.data.neuron_managers.icp.get_neuron_metrics(),
         }
     }
 
@@ -54,7 +62,9 @@ pub struct Metrics {
     pub canister_info: CanisterInfo,
     pub authorized_principals: Vec<Principal>,
     pub rewards_destination: Option<Principal>,
-    pub ogy_neuron_manager_metrics: Vec<NeuronWithMetric>,
+    pub ogy_neuron_manager_metrics: Vec<SnsNeuronWithMetric>,
+    pub goldao_neuron_manager_metrics: Vec<SnsNeuronWithMetric>,
+    pub icp_neuron_manager_metrics: Vec<NnsNeuronWithMetric>,
 }
 
 #[derive(CandidType, Deserialize, Serialize)]
@@ -77,18 +87,18 @@ pub struct Data {
 impl Data {
     pub fn new(
         authorized_principals: Vec<Principal>,
-        ogy_sns_governance_canister_id: CanisterId,
-        ogy_sns_ledger_canister_id: CanisterId,
-        ogy_sns_rewards_canister_id: CanisterId,
+        ogy_manager_config: OgyManagerConfig,
+        goldao_manager_config: GoldaoManagerConfig,
+        icp_manager_config: IcpManagerConfig,
         rewards_destination: Option<Principal>,
         now: TimestampMillis,
     ) -> Self {
         Self {
             authorized_principals,
             neuron_managers: NeuronManagers::init(
-                ogy_sns_governance_canister_id,
-                ogy_sns_ledger_canister_id,
-                ogy_sns_rewards_canister_id,
+                ogy_manager_config,
+                goldao_manager_config,
+                icp_manager_config,
                 now,
             ),
             rewards_destination,
@@ -100,29 +110,36 @@ impl Data {
 pub struct NeuronManagers {
     pub now: TimestampMillis,
     pub ogy: OgyManager,
+    pub goldao: GoldaoManager,
+    pub icp: IcpManager,
 }
 
 impl NeuronManagers {
     pub fn init(
-        ogy_sns_governance_canister_id: CanisterId,
-        ogy_sns_ledger_canister_id: CanisterId,
-        ogy_sns_rewards_canister_id: CanisterId,
+        ogy_manager_config: OgyManagerConfig,
+        goldao_manager_config: GoldaoManagerConfig,
+        icp_manager_config: IcpManagerConfig,
         now: TimestampMillis,
     ) -> Self {
         Self {
             now,
-            ogy: OgyManager {
-                ogy_sns_governance_canister_id,
-                ogy_sns_ledger_canister_id,
-                ogy_sns_rewards_canister_id,
-                neurons: Neurons::default(),
-            },
+            ogy: ogy_manager_config.into(),
+            goldao: goldao_manager_config.into(),
+            icp: icp_manager_config.into(),
         }
     }
 
     pub fn get_neurons(&self) -> NeuronList {
         NeuronList {
             ogy_neurons: self.ogy.neurons.all_neurons.clone(),
+            goldao_neurons: self.goldao.neurons.all_neurons.clone(),
+        }
+    }
+
+    pub fn get_neuron_manager(&self, neuron_type: NeuronType) -> NeuronManagerEnum {
+        match neuron_type {
+            NeuronType::OGY => NeuronManagerEnum::OgyManager(self.ogy.clone()),
+            NeuronType::GOLDAO => NeuronManagerEnum::GoldaoManager(self.goldao.clone()),
         }
     }
 }
@@ -130,4 +147,5 @@ impl NeuronManagers {
 #[derive(CandidType, Serialize)]
 pub struct NeuronList {
     ogy_neurons: Vec<Neuron>,
+    goldao_neurons: Vec<Neuron>,
 }