+++ backend/integration_testing/src/client/mod.rs
@@ -1,5 +1,8 @@
 pub mod icrc1;
 pub mod icrc1_icrc2_token;
 pub mod macros;
+pub mod nns_governance;
 pub mod pocket;
+pub mod sns_governance;
 pub mod sns_neuron_controller;
+pub mod sns_root_canister;