+++ backend/integration_testing/src/lib.rs
@@ -2,5 +2,6 @@
 
 mod client;
 pub mod sns_neuron_controller_suite;
+mod sns_test_env;
 mod utils;
 mod wasms;