+++ backend/integration_testing/src/sns_test_env/mod.rs
@@ -1,0 +1,3 @@
+pub mod nns_test_env;
+pub mod sns_init_args;
+pub mod sns_test_env;